Merge http://linuxusb.bkbits.net/linus-2.5 into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/07/24 10:39:13-07:00 petkan
[PATCH] USB: rtl8150 updated

  new vendor/device ID;
  redundant check removed from probe();

2002/07/24 09:48:37-07:00 greg
[PATCH] USB: changed the interface name to be a bit more unique.

This is needed as long as we have the directory of symlinks in the bus
subdir in driverfs to point to the unique interfaces.

2002/07/22 17:01:52-07:00 jb
[PATCH] drivers/usb/misc/tiglusb.c v1.04

Update to the latest version of the tiglusb driver

2002/07/22 16:52:18-07:00 greg
[PATCH] USB: usb-serial.h cleanups.

removed tty_driver from struct usb_serial, as it's not used.
(thanks to Stuart MacDonald for pointing this out.)

Updated the documentation for the structures to the proper style.

2002/07/22 16:23:55-07:00 greg
[PATCH] USB: rio500.c bugfix

patch from Silvio Cesare to fix signed problem.

2002/07/22 16:09:20-07:00 david-b
[PATCH] hid_ff_init could not find initializer...

Lately I get warnings whenever I plug in a device with
a HID interface ... when force feedback support is
configured, that's what it does.

Yeech ... that should should at most be a dbg().

2002/07/22 16:09:07-07:00 ch
[PATCH] set_device_description oops fixage mk2

Unlike previous version, this one doesn't oops and is perspicuous.
Please apply.

2002/07/22 16:08:54-07:00 ch
[PATCH] for ohci on SA-1111

This is needed by 2.5.26-rmk1.

2002/07/22 16:08:34-07:00 david-b
[PATCH] usb_set_interface() doc

This patch updates the kerneldoc for usb_set_interface()
to capture the point that it must not be called while
any URBs are pending to endpoints in that interface.
(Else state in the hcds, and devices, can get confused.)

2002/07/22 16:08:13-07:00 jaharkes
[PATCH] uhci-hcd suspend fix

Problem:

When resuming after a suspend, the driver disconnects. Later on, (next
usb interrupt?) or suspend/resume, the kernel oopses.

Cause:

hcd.state is never set to 'USB_STATE_READY' when resuming. The following
patch fixes this. Maybe Johannes might want to do it differently, i.e.
by adding hcd.state = USB_STATE_READY to uhci_resume instead of moving
the state change into start_hc.

2002/07/22 16:08:00-07:00 greg
[PATCH] USB pl2303: new device support added

Thanks to Ang Sei Heng <seiheng@talariax.com> for the information.

BKrev: 3d3f14f1EeQh4tYt7E0JTMARUuWBhw