[PATCH] Hot-plug CPU Boot Changes
This patch alters the boot sequence to "plug in" each CPU, one at a
time.  You need the patch for each architecture, as well.  The
interface used to be "smp_boot_cpus()", "smp_commence()", and each
arch implemented the "maxcpus" boot arg itself.  With this patch,
it is:
  smp_prepare_cpus(maxcpus): probe for cpus and set up cpu_possible(cpu).
  __cpu_up(cpu): called *after* initcalls, for each cpu where
     cpu_possible(cpu) is true.
  smp_cpus_done(maxcpus): called after every cpu has been brought up

BKrev: 3d41ccf7tMpR-8LtBqMcm9znDjQ_2g