[PATCH] sound/oss/trident.c [1/2] merge driver from 2.4-ac
This patch (1/2) brings the sound/oss/trident.c driver up to date with
the driver in the 2.4-ac tree. It fixes the following bugs:

* fix wrong cast in suspend/resume (Eric Lemar via Ian Soboroff)

* fix bug where we would free with free_pages() memory allocated via
pci_alloc_consistent().

* add a missing unlock on an error path.

* rewrite the code to read/write registers of audio codecs for Ali5451
(Lei Hu)

It also does various cleanups so that the code conforms to
Documentation/CodingStyle and is nicer to work with.

BKrev: 3d4480eee0CTYsmtDuqmHbA0dYEZeg