Merge bk://bk.arm.linux.org.uk:14691 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/07/29 15:53:04+01:00 rmk
[SERIAL] Cleanup includes.
Al Viro pointed out there was a fair bit of redundancy here.  We
remove many include files from the serial layer, leaving those
which are necessary for it to build.  This has been posted to lkml,
no one complained.

This cset also combines a missing include of asm/io.h in 8250_pci.c
(unfortunately I've lost the name of the reporter, sorry.)

2002/07/29 15:35:36+01:00 rmk
[SERIAL] Add HP Diva PCI serial port support.

2002/07/28 23:13:49+01:00 rmk
[SERIAL] Add pci_disable_device() to initialisation failure paths.

2002/07/28 23:01:26+01:00 rmk
[SERIAL] Remove some old compatibility cruft from 8250_pci.c
8250_pci.c contains some old compatibility cruft for when __devexit
wasn't defined by the generic kernel.  It is now, so it's gone.

2002/07/27 11:35:16+01:00 rmk
[SERIAL] Locking fixup (part 2)
After the last few days of debugging, we've ended up with the caller
of the start_tx and stop_tx methods taking the per-port lock.  This
cset and the accompanying csets make the same change to some of the
other methods for consistency reasons.  Since these methods don't
contain a lot of code, it is better that they have consistent locking
rules.

This cset fixes up the enable_ms method.

2002/07/27 11:27:42+01:00 rmk
[SERIAL] Locking fixup (part 1)
After the last few days of debugging, we've ended up with the caller
of the start_tx and stop_tx methods taking the per-port lock.  This
cset and the accompanying csets make the same change to some of the
other methods for consistency reasons.  Since these methods don't
contain a lot of code, it is better that they have consistent locking
rules.

This cset fixes up the stop_rx method.

BKrev: 3d455e54SjKkA8N1fZ58CBkl_ZmQlg