Split up "do_fork()" into "copy_process()" and "do_fork()".
copy_process() just copies the process, it doesn't actually start it.

This is in preparation for doing a "atomically start process on CPU X"
or other cases where we want to change the state of the process before
we actually start running it.

BKrev: 3d4717f3c548ZZVP-6cNg3DOMAXgfA