[PATCH] ohci, control/enumeration fix
One of the flakey things I see with 2.5.29 is that many
devices give trouble enumerating.  uhci still oopses khubd
enumerating pegasus (many control message errors first).
But that's not the only troublesome combo.

This patch fixes a failure that would hang the ohci driver
It's basically avoiding a few less IRQs (most noticable
with ISO), in this case so that a wedged SET_ADDRESS completes.
(Instead of hanging khubd in 'D' state.)  And it also updates
a diagnostic to provide more troubleshooting help for similar
issues in the future.

BKrev: 3d488579hBdZr7OYeQ9upFBBnlzrDA