Merge with Rusty 'trivial' Russell
2002/08/01 12:25:55-07:00 wstinson
[PATCH] small region change for boardergo.c

this is a small patch for boardergo.c linux driver for isdn HYSDN cards to
    1) remove call to check_region using request_region instead
    2) check the status of call to request_region
    3) release allocated region resource in case of error.

I don't have this hardware so compilation checked only.

2002/08/01 12:25:42-07:00 hch
[PATCH] documentation typos in

from Debian's kernel package

[ Typo fix, makes sense, please apply ]

2002/08/01 12:25:27-07:00 james
[PATCH] Typos in linux_arch_i386_kernel_io_apic.c


2002/08/01 12:25:17-07:00 gnb
[PATCH] PATCH 2.5: kconfig missing EXPERIMENTAL 4 (1_4)

Symbols CONFIG_HISAX_ENTERNOW_PCI and CONFIG_HISAX_AMD7930 depend on
CONFIG_EXPERIMENTAL but do not say so in their banners.

2002/08/01 12:25:07-07:00 gnb
[PATCH] PATCH 2.5: kconfig spurious EXPERIMENTAL 3 (1_2)

Symbol CONFIG_PPPOATM doesn't depend on CONFIG_EXPERIMENTAL but claims
to in its banner.

2002/08/01 12:24:57-07:00 maalanen
[PATCH] using ptr after kfree()

Somebody noticed on #kernelnewbies. In the error path, so this is
silly cleanup work...

2002/08/01 12:24:48-07:00 gnb
[PATCH] PATCH 2.5: kconfig spurious bool default value (1_3)

Some CML1 rule contributors seem to think that the "bool" statement
takes a third argument which is the default value.  It doesn't.

2002/08/01 12:24:40-07:00 rfjak
[PATCH] 2.5 Trivial patch - 1400x1050 video mode added twice in 2.5.28

I am not sure if you are the actual maintainer for this, but your name
adorns the source :
  2.5.28 introduced video mode 1400x1050 twice into drivers/video/modedb.c
Both instances use indentical timings, so I guess it's a mistake.

2002/08/01 12:24:30-07:00 gnb
[PATCH] PATCH 2.5: kconfig spurious EXPERIMENTAL 3 (2_2)

Symbol CONFIG_EXT3_FS doesn't depend on CONFIG_EXPERIMENTAL but claims
to in its banner.

2002/08/01 12:24:18-07:00 anton
[PATCH] fix warning

Quick fix for:
  smbiod.c:272: warning: deprecated use of label at end of compound statement

2002/08/01 12:24:03-07:00 wstinson
[PATCH] small region change for baycom_ser_hdx.c

this is a small patch for baycom_ser_hdx.c  -- baycom ser12 halfduplex radio modem driver to
    1) remove call to check_region using request_region instead
    2) check the status of call to request_region
    3) release allocated region resource in case of error.

I don't have this hardware so compilation checked only.

2002/08/01 12:23:49-07:00 hch
[PATCH] silence APIC errors a bit

They are harmless mostly and can happen in normal use. (from -rh)

2002/08/01 12:23:37-07:00 mikpe
[PATCH] ftape bitops -> long fix

Silence bitop on non-long argument warnings.

2002/08/01 12:23:29-07:00 james
[PATCH] Typo in linux_net_sched_sch_ingress.c


2002/08/01 12:23:21-07:00 hch
[PATCH] explicit signed char cast in i386 spin_is_locked

To make spinlock debugging work with -funsigned-char the i386 version
of spin_is_locked() needs to cast to signed char explicitly instead of
just char.  XFS needed -funsigned-char (and currently still has it),
but it doesn't evert hurt.

2002/08/01 12:23:13-07:00 ahaas
[PATCH] designated initializer patches for

Here are patches for designated initializers in kernel/*.

2002/08/01 12:23:05-07:00 hch
[PATCH] reparent scsi new-EH threads to init

(Included in 2.4)

2002/08/01 12:22:57-07:00 mikpe
[PATCH] drm_mga bitops -> long fix

Silence bitop on non-long argument warnings.

[ Not just silence warnings, but fix portability bug ]

2002/08/01 12:22:50-07:00 anton
[PATCH] fix compile warning

Fix warning about undefined struct pt_regs.

2002/08/01 12:22:42-07:00 trini
[PATCH] Export synchronize_irq on CONFIG_SMP=y

Currently, all arches which support SMP define synchronize_irq(irq)
to be a real function (generally defined in arch/$(ARCH)/kernel/irq.c).
This export was removed inadvertanly I believe in the "bit IRQ lock"
removal and IRQ cleanups ChangeSet.

2002/08/01 12:22:34-07:00 ica2_ts
[PATCH] Fix typo in net_ipv4_ipconfig.c

this just fixes a typo.

2002/08/01 12:22:27-07:00 ica2_ts
[PATCH] Fix typo in mm_slab.c

this just fixes a typo.

2002/08/01 12:22:19-07:00 rusty
[PATCH] Mark sparc32 unmaintained in 2.5

DaveM said this was fine.

2002/08/01 12:22:12-07:00 wstinson
[PATCH] region changes for rocket

this is a small patch for Rocketport device driver for Linux to
  1) remove two calls to check_region using request_region instead
  2) release allocated region resource in case of error.

I don't have this hardware so compilation checked only.

This patch does not remove all references to check_region in this driver
(one reference still left unchanged).

2002/08/01 12:22:05-07:00 mikpe
[PATCH] typo in fs_ufs_super.c:ufs_fill_super()

Obvious typo: checking block size but printing fragment size.

2002/08/01 12:21:58-07:00 gnb
[PATCH] PATCH 2.5: kconfig EXPERIMENTAL variant form

The convention is that symbols that depend on $CONFIG_EXPERIMENTAL are
marked in the banner with the string (EXPERIMENTAL).  There are some
cases where that string is correctly used but is Capitalised instead of
uppercase.

2002/08/01 12:21:47-07:00 gnb
[PATCH] PATCH: kconfig choice defaults 2 (3_3)

The default value specified as the last word of the "choice" statement
should be a unique abbreviation of one of the sub-prompts.  Using one
of the sub-symbols is not legal.  Fix 'Type of PHY'.

[Rusty: this is 3/3 partial resends of 'PATCH: kconfig choice defaults (1/2)']

2002/08/01 12:21:39-07:00 wli
[PATCH] remove declaration of __free_pte()

__free_pte() no longer exists in the kernel, but is still declared
from include/linux/mm.h. This patch removes that declaration.

2002/08/01 12:21:30-07:00 hch
[PATCH] update dqblk_xfs.h inclusion guards

This file was called xqm.h previously, and the inclusion guards still
think it is..

2002/08/01 12:21:22-07:00 colpatch
[PATCH] Adaptec Starfire config fix


BKrev: 3d498c8eWd5sJ9Ig3tvmn21T6lPD9g