Merge bk://ldm.bkbits.net/linux-2.5-driverfs-api into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/08/01 12:24:41-07:00 mochel
Merge osdl.org:/home/mochel/src/kernel/devel/linux-2.5-virgin
into osdl.org:/home/mochel/src/kernel/devel/linux-2.5-driverfs-api

2002/08/01 12:07:33-07:00 mochel
driverfs: define struct driverfs_ops and remove struct device dependencies

In order to read/write attributes, you have to deal directly with the object that owns them. driverfs really
wants to be generic and not deal directly with those objects. So, we create an intermediate layer that
subsystems must implement that converts between the generic objects and the specific objects that own the
attributes.

This allows allows attributes to be exported for any object type.

In doing so, it places the responsibility on the subsystems to do the following:

- define their own object-specific attribute structures
- define their own driverfs_ops
- set the ops pointer in struct driver_dir_entry when creating an object's directory
- do object reference counting on open() and close()
- call the show() and store() callbacks of their attribute structure
- convert between the generic objects and the specific objects from the struct driver_dir_entry and
  struct attribute pointers (using container_of)

The implementation of this layer for struct device is intended to be used as an example of the interface.

Because this layer of abstraction is now in place, we can move the device attribute structure into
include/linux/device.h, and driverfs should be free of references to it completely.


2002/08/01 11:17:11-07:00 mochel
driverfs: Add struct attribute
driverfs can only handle passing struct device to read/write functions. In order to free it of this limitation,
we need a common data structure for driverfs to pass around.

The only thing that driverfs really needs are the name and mode of the file, which are now located in struct
attribute.

struct device_attribute gets a struct attribute member, which holds the name and mode. With the DEVICE_ATTR
macro, users of the structure require no modification.

device_create_file is modified to take a struct attribute parameter

a to_dev_attr() macro is introduced to convert between a struct attribute to a struct device_attribute

2002/08/01 10:48:56-07:00 mochel
Convert users of struct device_attribute to initialize the structs using
DEVICE_ATTR macro.

2002/08/01 10:48:04-07:00 mochel
driverfs: Declare DEVICE_ATTR macro for initializing device attributes
(hide internal format of the structure)

2002/07/31 14:09:46-07:00 mochel
fixup users of device_remove_file to pass a struct device_attribute, instead of a char *

2002/07/31 14:08:54-07:00 mochel
driverfs:
make device_remove_file take a struct device_attribute *, instead of just
a char * (for consistency with device_create_file)

2002/07/30 20:54:58-07:00 mochel
driverfs: add device_remove_symlink wrapper for removing symlinks (since driverfs
API is just about to change...)

2002/07/30 20:44:14-07:00 mochel
Fixup users of driverfs:
do a s/driver_file_entry/device_attribute/g on all of them.

2002/07/30 20:43:13-07:00 mochel
driverfs: Change the name of struct driver_file_entry to struct device_attribute

It may seem gratuitous, but it's what we really want.

driverfs files are meant to expose attributes of various kernel objects, so in that sense,
the change adds more accurate meaning to the object.

Plus, we will soon gain the ability to expose attributes of drivers (both device and bus)
themselves, and we want to be able to have each mean something reasonable.

This changes driverfs and the device model core (but none of the other users)

BKrev: 3d499c54-LzRQPt-CyJUj4lHx_edFA