driverfs: decrement refcount on dentry being removed, not directory
This brain fart is left over from some cleanup of these functions a _long_
time ago. We need to dput() the dentry, since we have an implicit count of
one left over from the create function.

Instead, we were dput() on the directory that it was in, which we didn't
have a matching dget() for.

BKrev: 3d4ec034vxXmnFTQmpKI7wja9fI7-Q