[PATCH] PATCH - RPC/TCP 4 of 4 - Handle short read when reading RPC/TCP packet length
The code already copes with the possibility of having
read some of the 4byte length but not all, but can never
get into that situation.  We just need to add a 'return'
for the case where read got some data but not enough.
Either 0 or -EAGAIN could be returned, they both have the
same effect.

BKrev: 3d591c55lBEsC003fRMBWf04LIeIRQ