[PATCH] A generic RTC driver [1/3]
This is the current version of the genrtc driver from the m68k
community.  This is slightly different than the version I have sent
previously in that it has been switched to C99-style initializers,
which was done in the current m68k CVS tree by Geert Uytterhoeven, and
the needed changes to select/compile it in general.  I had previously
asked the m68k community if anyone objected to this being submitted by me,
and I got Richard Zidlicky's (who's at the top of the file) approval, as
well as Geert Uytterhoeven's approval.

BKrev: 3d59241b71apx4cC5OR1iC1SyM0cAw