Merge mulgrave.(none):/home/jejb/BK/scsi-cpqfc-2.5 into mulgrave.(none):/home/jejb/BK/scsi-for-linus-2.5
2002/08/03 20:03:19-05:00 steve.cameron
[PATCH] fix cpqfc passthrough ioctl for 2.5.30

This patch fixes the cpqfc passthrough ioctl for 2.5.30.

One question... Is it possible for a user app, by abusing
a upperlevel driver (e.g. the sg driver) to put a bogus
value into Scsi_Request->upper_level_private?

This patch kind of depends on this being impossible.

If it is possible, maybe it's better if I find another way
to identify passthrough commands on completion than what
this patch does.  e.g. keep a list of the passthru Scsi_Requests
that are outstanding, then search the list on each completion...(yuck.)

-- steve

2002/08/03 20:02:52-05:00 steve.cameron
[PATCH] cpqfc, 2.5.30, lun fix

This is basically a port of a patch to 2.4 kernels to fix
the ioctls to pay attention to the lun field.  For 2.5.30,
the passthrough ioctl is pretty broken, and this patch does not
fix that.  (The patch I'm sending after this one does.)

-- steve

BKrev: 3d5a70b7DSwVnlQMfiT60PJ3yr3nrQ