[PATCH] Thread exit notification by futex
This updates the CLONE_CLEARTID case to use futexes to make it easier
to wait for a thread exit.

glibc/pthreads had been updated to use the TID-futex, this removes an
extra system-call and it also simplifies the pthread_join() code.  The
pthreads testcode works just fine with the new kernel and does not work
with a kernel that does not do the futex wakeup, so it's working fine.

BKrev: 3d5e9fab_z3lLaCdcjGXXNBceyc_ng